refact : add back and front foleder and DDD layerd archtecture
