+++ src/main/java/christmas/back/domain/event/config/EventType.java
@@ -1,0 +1,9 @@
+package christmas.back.domain.event.config;
+
+public enum EventType {
+    DDayEvent,
+    SpecialEvent,
+    WeekendEvent,
+    WeekEvent,
+    GiftEvent
+}
