+++ src/main/java/christmas/back/infrastructure/repository/ClientRepositoryImpl.java
@@ -1,0 +1,16 @@
+package christmas.back.infrastructure.repository;
+
+import christmas.back.domain.user.ClientRepository;
+import christmas.back.domain.user.model.Client;
+
+public class ClientRepositoryImpl implements ClientRepository {
+    @Override
+    public Client save(Client client) {
+        return null;
+    }
+
+    @Override
+    public Client findById(Long id) {
+        return null;
+    }
+}
