--- src/main/java/christmas/model/event/config/BaseEvent.java
@@ -1,17 +1,0 @@
-package christmas.model.event.config;
-
-import christmas.model.User.Client;
-import java.util.Map;
-
-public class BaseEvent {
-    public Boolean canGetEvent(Client client) {
-        return null;
-    }
-    public Map<EventType,Integer> getEventBenefit(Client client) {
-        return null;
-    }
-
-    public void updateClientBenefit(Client client) {
-
-    }
-}
