--- src/main/java/christmas/model/event/config/EventType.java
@@ -1,9 +1,0 @@
-package christmas.model.event.config;
-
-public enum EventType {
-    DDayEvent,
-    SpecialEvent,
-    WeekendEvent,
-    WeekEvent,
-    GiftEvent
-}
