--- src/main/java/christmas/model/event/discount/DdayEvent.java
@@ -1,32 +1,0 @@
-package christmas.model.event.discount;
-
-import christmas.model.User.Client;
-import christmas.model.event.config.BaseEvent;
-import christmas.model.event.config.EventType;
-import java.util.HashMap;
-import java.util.Map;
-
-public class DdayEvent extends BaseEvent {
-    private final Integer day = 25;
-    @Override
-    public Boolean canGetEvent(Client client) {
-        if(client.getVisitDay() <= day) {
-            return true;
-        }
-        return false;
-    }
-    @Override
-    public Map<EventType,Integer> getEventBenefit(Client client) {
-        int amount = 1000 + (client.getVisitDay() - 1) * 100;
-        Map<EventType, Integer> benefitMap = new HashMap<>();
-        benefitMap.put(EventType.DDayEvent, amount);
-        return benefitMap;
-    }
-    @Override
-    public void updateClientBenefit(Client client) {
-        client.joinEvent();
-        int amount = 1000 + (client.getVisitDay() - 1) * 100;
-        client.addBenefitToTotalDiscountAmount(amount);
-        client.addBenefitToTotalEventAmount(amount);
-    }
-}
